outsource Logic
